% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_in_package_code.R
\name{find_in_package_code}
\alias{find_in_package_code}
\title{Find specific code snippets in packages.}
\usage{
find_in_package_code(pattern)
}
\arguments{
\item{pattern}{Same as in stringr::str_detect.}
}
\value{
\code{data.frame} with file and line where the pattern is found.
}
\description{
Find specific code snippets in packages.
}
\details{
Assumes that a .Rproj exists
}

