% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_in_package_code.R
\name{find_nonascii_in_package}
\alias{find_nonascii_in_package}
\title{Find non ASCII characters in packages.}
\usage{
find_nonascii_in_package(sub_path = "R")
}
\arguments{
\item{sub_path}{Path to look in.}

\item{pattern}{Same as in stringr::str_detect.}
}
\value{
\code{data.frame} with file and line where the pattern is found.
}
\description{
Find non ASCII characters in packages.
}
\details{
Assumes that a .Rproj exists
}
