% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_in_package_code.R
\name{parse_rproj}
\alias{parse_rproj}
\title{Function to parse Rproj files to R lists}
\usage{
parse_rproj(file_path)
}
\arguments{
\item{file_path}{Path to the .Rproj file}
}
\description{
Function to parse Rproj files to R lists
}
